Rename AuthController to LoginController and add it in auth folder
