+++ .eslintrc.json
@@ -26,7 +26,7 @@
         ],
         "linebreak-style": [
             "error",
-            "windows"
+            "unix"
         ]
         
     },
