+++ client/.eslintrc.json
@@ -30,7 +30,7 @@
         ],
         "linebreak-style": [
             "error",
-            "windows"
+            "unix"
         ]
     }
 }
