+++ server/controllers/index.ts
@@ -1,5 +1,5 @@
 /* eslint-disable import/prefer-default-export */
-import login from './authController';
+import login from './auth/loginController';
 
 export {
   login,
